Refactor GQL schema definition
