+++ package.json
@@ -20,6 +20,7 @@
     "i18next-browser-languagedetector": "^4.0.1",
     "i18next-sprintf-postprocessor": "^0.2.2",
     "immutable": "^4.0.0-rc.12",
+    "lodash": "^4.17.15",
     "react": "^16.11.0",
     "react-apollo": "^3.1.3",
     "react-dom": "^16.11.0",
