+++ server/index.js
@@ -1,8 +1,7 @@
-const { ApolloServer } = require('apollo-server');
-const typeDefs = require('./graphql/typeDefs');
-const resolvers = require('./graphql/resolvers');
+const { ApolloServer, makeExecutableSchema } = require('apollo-server');
+const schema = require('./graphql/schema');
 
-const server = new ApolloServer({ typeDefs, resolvers });
+const server = new ApolloServer({ schema });
 
 server.listen().then(({ url }) => {
   console.log(`🚀  Server ready at ${url}`);
